separate gulpfile and change strict rule into eslintrc for tests.
