+++ .eslintrc
@@ -10,7 +10,7 @@
   },
   "rules": {
     // Strict mode
-    "strict": [2, "global"],
+    "strict": 0,
 
     // Code style
     "indent": [2, 2],
