+++ gulp/globalVars.js
@@ -1,0 +1,10 @@
+'use strict';
+
+//it's just a simple singletone object for connecting different modules
+//maybe it's not a good solution and we need to organize communication
+//another way
+
+module.exports = {
+  watch: false,
+  browserSync: null
+};
