+++ gulp/tasks/arguments.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('minimist')(process.argv.slice(2));
