+++ gulp/tasks/deploy.js
@@ -1,0 +1,28 @@
+'use strict';
+
+var gulp = require('gulp');
+var argv = require('./arguments');
+var path = require('path');
+var $ = require('../util/load-plugin');
+var del = require('del');
+
+
+// Deploy to GitHub Pages
+gulp.task('deploy', function() {
+
+  // Remove temp folder
+  if (argv.clean) {
+    var os = require('os');
+    var repoPath = path.join(os.tmpdir(), 'tmpRepo');
+    $.util.log('Delete ' + $.util.colors.magenta(repoPath));
+    del.sync(repoPath, {force: true});
+  }
+
+  return gulp.src('build/**/*')
+    .pipe($.if('**/robots.txt', !argv.production ?
+      $.replace('Disallow:', 'Disallow: /') : $.util.noop()))
+    .pipe($.ghPages({
+      remoteUrl: 'https://github.com/{name}/{name}.github.io.git',
+      branch: 'master'
+    }));
+});
