+++ gulp/util/load-plugin.js
@@ -1,0 +1,4 @@
+'use strict';
+
+var $ = require('gulp-load-plugins')();
+module.exports = $;
