+++ package.json
@@ -59,6 +59,7 @@
     "protractor": "^2.0.0",
     "psi": "^1.0.6",
     "react-tools": "^0.13.1",
+    "require-dir": "^0.3.0",
     "run-sequence": "^1.0.2",
     "style-loader": "^0.10.1",
     "url-loader": "^0.5.5",
